[ninja][location] match other for this auth.
given one encrypted location id from contamination log, location auth would
fetch other entries in contamination log at the same location at the same time.
it needs to decrypt (and not just query), because of padding. this is in
addition to the exisiting logic to forward to other location auths.

more logging and fine-grain, synchronized timestamps.